# Tiny Set optimizations