--- a/set/dict.go
+++ b/set/dict.go
@@ -29,7 +29,7 @@ var pool = sync.Pool{}
 
 // Set is an implementation of ISet using the builtin map type. Set is threadsafe.
 type Set struct {
-	items     map[interface{}]bool
+	items     map[interface{}]struct{}
 	lock      sync.RWMutex
 	flattened []interface{}
 }
@@ -41,7 +41,7 @@ func (set *Set) Add(items ...interface{}) {
 
 	set.flattened = nil
 	for _, item := range items {
-		set.items[item] = true
+		set.items[item] = struct{}{}
 	}
 }
 
@@ -94,7 +94,7 @@ func (set *Set) Clear() {
 	set.lock.Lock()
 	defer set.lock.Unlock()
 
-	set.items = map[interface{}]bool{}
+	set.items = map[interface{}]struct{}{}
 }
 
 // All returns a bool indicating if all of the supplied items exist in the set.
@@ -134,7 +134,7 @@ func (set *Set) Dispose() {
 func New(items ...interface{}) *Set {
 	set := pool.Get().(*Set)
 	for _, item := range items {
-		set.items[item] = true
+		set.items[item] = struct{}{}
 	}
 
 	return set
@@ -143,7 +143,7 @@ func New(items ...interface{}) *Set {
 func init() {
 	pool.New = func() interface{} {
 		return &Set{
-			items: make(map[interface{}]bool, 10),
+			items: make(map[interface{}]struct{}, 10),
 		}
 	}
 }