1. `bool` is replaced with `struct{}` to save some memory in large sets
1. Use direct call of (R)Unlock in really plain simple methods like `Exists`, `Len`, `Clear` to speedup Set

with `defer()`
```
BenchmarkLen	50000000	        52.8 ns/op
BenchmarkExists	20000000	        93.4 ns/op
BenchmarkClear	10000000	       215 ns/op
```
without `defer()`
```
BenchmarkLen	100000000	        21.9 ns/op
BenchmarkExists	50000000	        51.5 ns/op
BenchmarkClear	10000000	       167 ns/op
```