@cep21 2015-02-11
The defer optimization is somewhat surprising for me.  My only guess on why it's not so easy to optimize the same way is that defer() will also do the unlock if something inside panics, so if the code ever panics because of some bug users of the data structure would be unable to recover.  Any other thoughts?