@noxiouz 2015-02-11
I can show you a lot of places in stdlib where defer is not used to unlock. On the other side I replaced defer only in functions without possibilities to panic and one point to return. It's hard to expect a situation where ``` _, ok := m["a"]``` would panic. It's a really simple cases. Also there's some kind of research here: http://lk4d4.darth.io/posts/defer/