@alexandercampbell-wf 2015-02-11
Theoretically you could recover from a panic and continue using the datastructure, but whatever caused the panic probably left it in a bad state. I wouldn't trust the data at that point.

Defer is very convenient and aids safety in larger functions with multiple return sites, but I agree that in simple functions like these, the performance improvement is worth explicitly unlocking before the return.